LPS-63097 Move portal-soap-extender-test to portal-remote-soap-extender-test
