+++ modules/portal-remote/portal-remote-soap-extender-test/bnd-api-handler.bnd
@@ -1,0 +1,10 @@
+Bundle-Activator: com.liferay.portal.remote.soap.extender.test.activator.handler.JaxWsApiBundleActivator
+Bundle-Name: Liferay Portal Remote SOAP Extender Test API Handler
+Bundle-SymbolicName: com.liferay.portal.remote.soap.extender.test.api.handler
+Bundle-Version: 1.0.0
+Include-Resource:\
+	com/liferay/portal/remote/soap/extender/test/activator/configuration=classes/com/liferay/portal/remote/soap/extender/test/activator/configuration,\
+	com/liferay/portal/remote/soap/extender/test/activator/handler=classes/com/liferay/portal/remote/soap/extender/test/activator/handler,\
+	com/liferay/portal/remote/soap/extender/test/handler=classes/com/liferay/portal/remote/soap/extender/test/handler,\
+	com/liferay/portal/remote/soap/extender/test/service=classes/com/liferay/portal/remote/soap/extender/test/service,\
+	com/liferay/portal/remote/soap/extender/test/util=classes/com/liferay/portal/remote/soap/extender/test/util
