+++ modules/portal-remote/portal-remote-soap-extender-test/bnd.bnd
@@ -1,0 +1,6 @@
+Bundle-Name: Liferay Portal Remote SOAP Extender Test
+Bundle-SymbolicName: com.liferay.portal.remote.soap.extender.test
+Bundle-Version: 1.0.0
+Liferay-Releng-Module-Group-Description:
+Liferay-Releng-Module-Group-Title: Remote Service Engines
+-include: ../../../marketplace/platform/bnd.bnd
