--- modules/portal/portal-soap-extender-test/bnd.bnd
@@ -1,6 +1,0 @@
-Bundle-Name: Liferay Portal SOAP Extender Test
-Bundle-SymbolicName: com.liferay.portal.soap.extender.test
-Bundle-Version: 1.0.0
-Liferay-Releng-Module-Group-Description:
-Liferay-Releng-Module-Group-Title: Remote Service Engines
--include: ../../../marketplace/platform/bnd.bnd
