--- modules/portal/portal-soap-extender-test/src/testIntegration/java/com/liferay/portal/soap/extender/test/JaxWsApiRegistrationTest.java
@@ -1,60 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portal.soap.extender.test;
-
-import com.liferay.portal.soap.extender.test.service.Greeter;
-
-import java.net.URL;
-
-import javax.xml.namespace.QName;
-import javax.xml.ws.Service;
-
-import org.arquillian.liferay.deploymentscenario.annotations.BndFile;
-
-import org.jboss.arquillian.container.test.api.RunAsClient;
-import org.jboss.arquillian.junit.Arquillian;
-import org.jboss.arquillian.test.api.ArquillianResource;
-
-import org.junit.Assert;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-/**
- * @author Carlos Sierra Andrés
- */
-@BndFile("bnd-api.bnd")
-@RunAsClient
-@RunWith(Arquillian.class)
-public class JaxWsApiRegistrationTest {
-
-	@Test
-	public void testGreeter() throws Exception {
-		URL url = new URL(_url, "/o/soap-test/greeterApi?wsdl");
-
-		QName qName = new QName(
-			"http://service.test.extender.soap.portal.liferay.com/",
-			"GreeterImplService");
-
-		Service service = Service.create(url, qName);
-
-		Greeter greeter = service.getPort(Greeter.class);
-
-		Assert.assertEquals("Greetings.", greeter.greet());
-	}
-
-	@ArquillianResource
-	private URL _url;
-
-}
